+++ packages/core/tests/js/imports/remove-import/input.ts
@@ -1,0 +1,2 @@
+import { n1, n2 } from 'p1';
+import { n3 } from 'p3';
