+++ packages/core/tests/js/imports/remove-import/output.ts
@@ -1,0 +1,1 @@
+import { n1 } from 'p1';
