+++ packages/core/tests/js/imports/remove-import/run.ts
@@ -1,0 +1,6 @@
+import { imports, type AstTypes } from '@sveltejs/cli-core/js';
+
+export function run(ast: AstTypes.Program): void {
+	imports.remove(ast, { name: 'n2', from: 'p1' });
+	imports.remove(ast, { name: 'n3', from: 'p3' });
+}
