+++ packages/core/tooling/js/imports.ts
@@ -146,7 +146,7 @@ function addImportIfNecessary(
 
 export function find(
 	ast: AstTypes.Program,
-	options: { name: string }
+	options: { name: string; from: string }
 ):
 	| { statement: AstTypes.ImportDeclaration; alias: string }
 	| { statement: undefined; alias: undefined } {
@@ -155,7 +155,7 @@ export function find(
 
 	Walker.walk(ast as AstTypes.Node, null, {
 		ImportDeclaration(node) {
-			if (node.specifiers) {
+			if (node.specifiers && node.source.value === options.from) {
 				const specifier = node.specifiers.find(
 					(sp) =>
 						sp.type === 'ImportSpecifier' &&
@@ -176,4 +176,36 @@ export function find(
 	}
 
 	return { statement: undefined, alias: undefined };
+}
+
+export function remove(
+	ast: AstTypes.Program,
+	options: {
+		name: string;
+		from: string;
+		statement?: AstTypes.ImportDeclaration; // Just in case you want to pass the statement directly
+	}
+): void {
+	const statement =
+		options.statement ?? find(ast, { name: options.name, from: options.from }).statement;
+
+	if (!statement) {
+		return;
+	}
+
+	if (statement.specifiers?.length === 1) {
+		const idxToRemove = ast.body.indexOf(statement);
+		ast.body.splice(idxToRemove, 1);
+	} else {
+		// otherwise, just remove the `defineConfig` specifier
+		const idxToRemove = statement.specifiers?.findIndex(
+			(s) =>
+				s.type === 'ImportSpecifier' &&
+				s.imported.type === 'Identifier' &&
+				s.imported.name === options.name
+		);
+		if (idxToRemove !== undefined && idxToRemove !== -1) {
+			statement.specifiers?.splice(idxToRemove, 1);
+		}
+	}
 }
